moved some enums & interfaces to types.ts
